refactoring inversion solution class;
